moving useTipReason to api module
